+++ enterprise/frontend/src/metabase-enterprise/data-studio/metrics/pages/MetricOverviewPage/CardOverview/CardOverview.tsx
@@ -1,8 +1,7 @@
 import { EntityCreationInfo } from "metabase/common/components/EntityCreationInfo";
 import { Flex, Stack } from "metabase/ui";
+import { OverviewVisualization } from "metabase-enterprise/data-studio/common/components/OverviewVisualization";
 import type { Card } from "metabase-types/api";
-
-import { OverviewVisualization } from "../OverviewVisualization";
 
 import S from "./CardOverview.module.css";
 import { DescriptionSection } from "./DescriptionSection";
