+++ enterprise/frontend/src/metabase-enterprise/data-studio/metrics/pages/MetricOverviewPage/CardOverview/DescriptionSection/DescriptionSection.tsx
